+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The version numbers are structured like `GENERATION.MAJOR.MINOR.PATCH`:
 - Add composer dependency `ext-json: *` as JSON requests and responses needs to be sent and parsed
 - Add contract `\Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\Contract\ApiConfigurationStorageInterface` to provide `\Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\ApiConfiguration`, that is used to identify and authenticate against a Shopware API
 - Add utility `\Heptacom\HeptaConnect\Package\Shopware6\Support\JsonStreamUtility` to unify JSON handling
+- Add utility `\Heptacom\HeptaConnect\Package\Shopware6\Support\LetterCase` to convert strings between different letter cases
 - Add exception code `1680371700` to `\Heptacom\HeptaConnect\Package\Shopware6\Support\JsonStreamUtility::fromPayloadToStream` when non-JSON-specific encoding issues occur
 - Add exception code `1680371701` to `\Heptacom\HeptaConnect\Package\Shopware6\Support\JsonStreamUtility::fromStreamToPayload` when the decoded JSON is not a PHP array
 - Add exception `\Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\Exception\AuthenticationFailed` to identify issues on authenticating with the Shopware API